fix: update swagger issues and update generation flow
Fixes #33
Adds block execute to definition. This should be finally fixed upstream
as well.

Fixes #35
Updates endpoint response to be of type Object. This one is interesting
because it seems like it should actually be returning a binary object
but instead it returns either a CSV string or JSON object depending on
the type of scheduled report that was chosen. When this is updated
upstream we will treat this as a binary download such as downloading a
sensor.

This also fixes dealing with text/* returns such as csv. Now we just
return it as a string instead of failing.

The Makefile has been updated to make use of the transformation.jq file
to ensure our swagger spec file has fixes needed while we wait for
upstream support.